+++ Detector/LinkDef.h
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 // SPDX-FileCopyrightText: Copyright CERN for the benefit of the SHiP Collaboration
 
+// $Id: DetectorLinkDef.h
+
 #ifdef __CINT__
 
 #pragma link off all globals;
@@ -8,8 +10,6 @@
 #pragma link off all functions;
 
 #pragma link C++ namespace SHiP;
-#pragma link C++ class TimeDet+;
-#pragma link C++ class TimeDetPoint+;
-#pragma link C++ class TimeDetHit+;
+#pragma link C++ class SHiP::DetectorPoint+;
 
 #endif