+++ TimeDet/TimeDet.h
@@ -72,9 +72,10 @@ class TimeDet : public FairDetector, public ISTLPointContainer {
   /**      This method is an example of how to add your own point
    *       of type TimeDetPoint to the clones array
    */
-  TimeDetPoint* AddHit(Int_t trackID, Int_t detID, TVector3 pos, TVector3 mom,
-                       Double_t time, Double_t length, Double_t eLoss,
-                       Int_t pdgCode, TVector3 Lpos, TVector3 Lmom);
+  TimeDetPoint* AddHit(Int_t eventID, Int_t trackID, Int_t detID, TVector3 pos,
+                       TVector3 mom, Double_t time, Double_t length,
+                       Double_t eLoss, Int_t pdgCode, TVector3 Lpos,
+                       TVector3 Lmom);
 
   virtual void EndOfEvent();
   virtual void FinishPrimary() { ; }
@@ -86,6 +87,7 @@ class TimeDet : public FairDetector, public ISTLPointContainer {
 
  private:
   /** Track information to be stored until the track leaves the active volume.*/
+  Int_t fEventID;       //!  event index
   Int_t fTrackID;       //!  track index
   Int_t fVolumeID;      //!  volume id
   TLorentzVector fPos;  //!  position at entrance
@@ -126,7 +128,7 @@ class TimeDet : public FairDetector, public ISTLPointContainer {
   TimeDet& operator=(const TimeDet&);
   Int_t InitMedium(const char* name);
 
-  ClassDef(TimeDet, 3)
+  ClassDef(TimeDet, 4)
 };
 
 #endif  // TIMEDET_TIMEDET_H_