+++ strawtubes/strawtubesPoint.h
@@ -5,16 +5,17 @@
 #ifndef STRAWTUBES_STRAWTUBESPOINT_H_
 #define STRAWTUBES_STRAWTUBESPOINT_H_
 
-#include "FairMCPoint.h"
+#include "DetectorPoint.h"
 #include "TObject.h"
 #include "TVector3.h"
 
 class strawtubesHit;
 
-class strawtubesPoint : public FairMCPoint {
+class strawtubesPoint : public SHiP::DetectorPoint {
  public:
   /** Default constructor **/
   strawtubesPoint();
+  using SHiP::DetectorPoint::DetectorPoint;
 
   /** Constructor with arguments
    *@param trackID  Index of MCTrack
@@ -25,27 +26,23 @@ class strawtubesPoint : public FairMCPoint {
    *@param length   Track length since creation [cm]
    *@param eLoss    Energy deposit [GeV]
    **/
-  strawtubesPoint(Int_t trackID, Int_t detID, TVector3 pos, TVector3 mom,
-                  Double_t tof, Double_t length, Double_t eLoss, Int_t pdgcode,
-                  Double_t dist);
+  strawtubesPoint(Int_t eventID, Int_t trackID, Int_t detID, TVector3 pos,
+                  TVector3 mom, Double_t tof, Double_t length, Double_t eLoss,
+                  Int_t pdgcode, Double_t dist);
 
   /** Destructor **/
   virtual ~strawtubesPoint();
 
-  /** Copy constructor **/
-  strawtubesPoint(const strawtubesPoint& point) = default;
-  strawtubesPoint& operator=(const strawtubesPoint& point) = default;
-
   /** Output to screen **/
-  using FairMCPoint::Print;
-  virtual void Print() const;
-  Int_t PdgCode() const { return fPdgCode; }
+
   Double_t dist2Wire() const { return fdist2Wire; }
 
+  void setDetectorName() override;
+  void extraPrintInfo() const override;
+
  private:
-  Int_t fPdgCode;
   Double_t fdist2Wire;
-  ClassDef(strawtubesPoint, 4);
+  ClassDefOverride(strawtubesPoint, 5);
 };
 
 #endif  // STRAWTUBES_STRAWTUBESPOINT_H_