+++ veto/veto.h
@@ -91,9 +91,10 @@ class veto : public FairDetector, public ISTLPointContainer {
   /**      This method is an example of how to add your own point
    *       of type vetoPoint to the clones array
    */
-  vetoPoint* AddHit(Int_t trackID, Int_t detID, TVector3 pos, TVector3 mom,
-                    Double_t time, Double_t length, Double_t eLoss,
-                    Int_t pdgcode, TVector3 Lpos, TVector3 Lmom);
+  vetoPoint* AddHit(Int_t eventID, Int_t trackID, Int_t detID, TVector3 pos,
+                    TVector3 mom, Double_t time, Double_t length,
+                    Double_t eLoss, Int_t pdgcode, TVector3 Lpos,
+                    TVector3 Lmom);
 
   /** The following methods can be implemented if you need to make
    *  any optional action in your detector during the transport.
@@ -121,6 +122,8 @@ class veto : public FairDetector, public ISTLPointContainer {
   /** Track information to be stored until the track leaves the
   active volume.
   */
+  //!  event index
+  Int_t fEventID;
   //!  track index
   Int_t fTrackID;
   //!  volume id
@@ -250,7 +253,7 @@ class veto : public FairDetector, public ISTLPointContainer {
 
   TGeoVolume* MakeSegments();
 
-  ClassDef(veto, 1)
+  ClassDef(veto, 2)
 };
 
 #endif  // VETO_VETO_H_