+++ veto/vetoPoint.h
@@ -7,15 +7,16 @@
 
 #include <array>
 
-#include "FairMCPoint.h"
+#include "DetectorPoint.h"
 #include "TObject.h"
 #include "TVector3.h"
 
-class vetoPoint : public FairMCPoint {
+class vetoPoint : public SHiP::DetectorPoint {
  public:
   /** Default constructor **/
   vetoPoint();
 
+  using SHiP::DetectorPoint::DetectorPoint;
   /** Constructor with arguments
    *@param trackID  Index of MCTrack
    *@param detID    Detector ID
@@ -28,29 +29,17 @@ class vetoPoint : public FairMCPoint {
    *@param length   Track length since creation [cm]
    *@param eLoss    Energy deposit [GeV]
    **/
-  vetoPoint(Int_t trackID, Int_t detID, TVector3 pos, TVector3 mom,
-            Double_t tof, Double_t length, Double_t eLoss, Int_t pdgCode,
-            TVector3 Lpos, TVector3 Lmom, Int_t eventID = -1);
 
   /** Destructor **/
   virtual ~vetoPoint();
 
-  /** Copy constructor **/
-  vetoPoint(const vetoPoint& point) = default;
-  vetoPoint& operator=(const vetoPoint& point) = default;
+  void setDetectorName() override;
+  void extraPrintInfo() const override;
 
   /** Output to screen **/
-  using FairMCPoint::Print;
-  virtual void Print() const;
-  Int_t PdgCode() const { return fPdgCode; }
-  TVector3 LastPoint() const { return TVector3(fLpos[0], fLpos[1], fLpos[2]); }
-  TVector3 LastMom() const { return TVector3(fLmom[0], fLmom[1], fLmom[2]); }
 
  private:
-  Int_t fPdgCode;
-  std::array<Double_t, 3> fLpos, fLmom;
-
-  ClassDef(vetoPoint, 5)
+  ClassDefOverride(vetoPoint, 5)
 };
 
 #endif  // VETO_VETOPOINT_H_